chore: Move RLS to src/pages folder
